requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
websockets==14.2
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
xgboost==2.1.1
scikit-learn==1.5.1
shap==0.45.1
google-genai==1.5.0
pydantic==2.9.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
authlib==1.4.0
httpx==0.28.1
groq==1.0.0
python-dotenv==1.0.1
itsdangerous==2.2.0
